Dev marshal as 4130 latest

+++ web/src/components/FOI/Home/Redlining.js
@@ -711,29 +711,34 @@ const Redlining = React.forwardRef(
             if (action === "delete") {
               let annotObjs = [];
               for (let annot of annots[0].children) {
-                let displayedDoc =
-                  pageMappedDocs.stitchedPageLookup[
-                    Number(annot.attributes.page) + 1
-                  ];
-                let individualPageNo = displayedDoc.page;
-                if (annot.name === "redact") {
-                  annotObjs.push({
-                    page: annot.attributes.page,
-                    name: annot.attributes.name,
-                    type: annot.name,
-                  });
-                } else {
-                  deleteAnnotation(
-                    requestid,
-                    displayedDoc.docid,
-                    displayedDoc.docversion,
-                    currentLayer.redactionlayerid,
-                    annot.attributes.name,
-                    (data) => {},
-                    (error) => {
-                      console.log(error);
-                    }
-                  );
+                let customData = annot.children.find(
+                  (element) => element.name == "trn-custom-data"
+                );
+                if (!customData?.attributes?.bytes?.includes("isDelete")){
+                  let displayedDoc =
+                    pageMappedDocs.stitchedPageLookup[
+                      Number(annot.attributes.page) + 1
+                    ];
+                  let individualPageNo = displayedDoc.page;
+                  if (annot.name === "redact") {
+                    annotObjs.push({
+                      page: annot.attributes.page,
+                      name: annot.attributes.name,
+                      type: annot.name,
+                    });
+                  } else {
+                    deleteAnnotation(
+                      requestid,
+                      displayedDoc.docid,
+                      displayedDoc.docversion,
+                      currentLayer.redactionlayerid,
+                      annot.attributes.name,
+                      (data) => {},
+                      (error) => {
+                        console.log(error);
+                      }
+                    );
+                  }
                 }
               }
               setDeleteQueue(annotObjs);
@@ -754,6 +759,34 @@ const Redlining = React.forwardRef(
                       flagid: pageFlagTypes["Withheld in Full"],
                       docid: displayedDoc.docid,
                     });
+
+                    let parentRedaction;
+                    let allAnnotations =
+                      docInstance.Core.annotationManager.getAnnotationsList();
+                    let _selectedAnnotations = allAnnotations.find(
+                      (annot) =>
+                        annot.Subject == "Free Text" &&
+                        annot.getCustomData("trn-redaction-type") ==
+                          "fullPage" &&
+                        annot.PageNumber == Number(annotatn.attributes.page) + 1
+                    );
+                    if (!!_selectedAnnotations) {
+                      parentRedaction = allAnnotations?.find(
+                        (r) =>
+                          r.Subject === "Redact" &&
+                          r.InReplyTo === _selectedAnnotations?.Id
+                      );
+                    }
+                    if (!!parentRedaction) {
+                      annotations[i].setCustomData(
+                        "existingId",
+                        parentRedaction?.Id
+                      );
+                      annotations[i].setCustomData(
+                        "existingFreeTextId",
+                        _selectedAnnotations?.Id
+                      );
+                    }
                   } else {
                     pageSelectionList.push({
                       page: Number(individualPageNo),
@@ -770,6 +803,7 @@ const Redlining = React.forwardRef(
                     "redactionlayerid",
                     currentLayer.redactionlayerid
                   );
+                  annotations[i].IsHoverable = false;
                 });
                 setPageSelections(pageSelectionList);
                 let annot = annots[0].children[0];
@@ -780,7 +814,7 @@ const Redlining = React.forwardRef(
                   });
                 setNewRedaction({
                   pages: annot.attributes.page,
-                  name: annot.attributes.name,
+                  names: annots[0].children.map((a) => a.attributes.name),
                   astr: astr,
                   type: annot.name,
                 });
@@ -824,7 +858,14 @@ const Redlining = React.forwardRef(
                 );
               }
             } else if (action === "modify") {
-              // handles saving modify actions and initial save of redaction due to grouping
+              console.log(info);
+              if (
+                info.source === "group" &&
+                newRedaction.astr.includes(annotations[0].Id) // if we are grouping the newly created annotations do not save
+              ) {
+                return;
+              }
+              // handles saving modify actions
               let selectedAnnotations =
                 docInstance.Core.annotationManager.getSelectedAnnotations();
               let username = docViewer
@@ -847,7 +888,6 @@ const Redlining = React.forwardRef(
                     requestid,
                     astr,
                     (data) => {
-                      setNewRedaction(null);
                       fetchPageFlag(
                         requestid,
                         currentLayer.redactionlayerid,
@@ -858,12 +898,7 @@ const Redlining = React.forwardRef(
                       console.log(error);
                     },
                     currentLayer.redactionlayerid,
-                    newRedaction
-                      ? createPageFlagPayload(
-                          pageSelections,
-                          currentLayer.redactionlayerid
-                        )
-                      : null
+                    null
                   );
                   const _resizeAnnot = {
                     pages: annot.attributes.page,
@@ -886,7 +921,6 @@ const Redlining = React.forwardRef(
                     requestid,
                     astr,
                     (data) => {
-                      setNewRedaction(null);
                       fetchPageFlag(
                         requestid,
                         currentLayer.redactionlayerid,
@@ -897,12 +931,7 @@ const Redlining = React.forwardRef(
                       console.log(error);
                     },
                     currentLayer.redactionlayerid,
-                    newRedaction
-                      ? createPageFlagPayload(
-                          pageSelections,
-                          currentLayer.redactionlayerid
-                        )
-                      : null
+                    null
                   );
                 }
               }
@@ -1182,6 +1211,9 @@ const Redlining = React.forwardRef(
         xml = parser.toString(xml);
         const _annotations = await annotManager.importAnnotations(xml);
         _annotations.forEach((_annotation) => {
+          if (_annotation.Subject === "Redact") {
+            _annotation.IsHoverable = false;
+          }
           annotManager.redrawAnnotation(_annotation);
           annotManager.setPermissionCheckCallback((author, _annotation) => {
             if (_annotation.Subject !== "Redact" && author !== username) {
@@ -1309,7 +1341,7 @@ const Redlining = React.forwardRef(
       else return _resizeAnnot;
     };
 
-    const saveRedaction = (_resizeAnnot = {}) => {
+    const saveRedaction = async (_resizeAnnot = {}) => {
       setModalOpen(false);
       setSaveDisabled(true);
       let redactionObj = getRedactionObj(newRedaction, editAnnot, _resizeAnnot); //newRedaction? newRedaction:  (editAnnot ? editAnnot :_resizeAnnot);
@@ -1409,7 +1441,10 @@ const Redlining = React.forwardRef(
         }
         // add section annotation
         var sectionAnnotations = [];
+        let annotationsToDelete = [];
         for (const node of astr.getElementsByTagName("annots")[0].children) {
+          let annotationsToDelete = [];
+
           let redaction = annotManager.getAnnotationById(node.attributes.name);
           let coords = node.attributes.coords;
           let X = coords?.substring(0, coords.indexOf(","));
@@ -1440,9 +1475,25 @@ const Redlining = React.forwardRef(
           const pageMatrix = doc.getPageMatrix(annot.PageNumber);
           const pageRotation = doc.getPageRotation(annot.PageNumber);
           annot.fitText(pageInfo, pageMatrix, pageRotation);
-          if (redaction.type == "fullPage")
+          if (redaction.type == "fullPage") {
             annot.setCustomData("trn-redaction-type", "fullPage");
-
+            let txt = new DOMParser().parseFromString(
+              node.getElementsByTagName("trn-custom-data")[0].attributes.bytes,
+              "text/html"
+            );
+            let customData = JSON.parse(txt.documentElement.textContent);
+            annot.setCustomData("existingId", customData.existingFreeTextId);
+            //Setting the existing annotationId in the new annotations for deleting
+            //from backend.
+            let existingFreeTextAnnot= annotManager.getAnnotationById(customData.existingFreeTextId);
+            let existingRedactAnnot= annotManager.getAnnotationById(customData.existingId);
+            if(!!existingFreeTextAnnot && !!existingRedactAnnot){
+              existingFreeTextAnnot.setCustomData("isDelete", true);
+              existingRedactAnnot.setCustomData("isDelete", true);
+              annotationsToDelete.push(existingFreeTextAnnot);
+              annotationsToDelete.push(existingRedactAnnot);
+            }
+          }
           sectionAnnotations.push(annot);
           redactionInfo.push({
             annotationname: redactionObj.name,
@@ -1451,12 +1502,48 @@ const Redlining = React.forwardRef(
           for (let section of redactionSections) {
             section.count++;
           }
+          //delete if there are existing fullpage redactions
+          if(annotationsToDelete?.length > 0){
+            annotManager.deleteAnnotations(annotationsToDelete, {
+              force: true,
+            });
+          }
           // grouping the section annotation with the redaction will trigger a modify event, which will also save the redaction
-          annotManager.groupAnnotations(annot, [redaction]);
+          await annotManager.groupAnnotations(annot, [redaction]);
         }
+        // annotManager.deleteAnnotations(annotationsToDelete, {
+        //     force: true,
+        //   });
+        let annotationList = [];
+        for (let name of newRedaction.names) {
+          annotationList.push(annotManager.getAnnotationById(name));
+        }
+        astr = await annotManager.exportAnnotations({
+          annotList: annotationList,
+          useDisplayAuthor: true,
+        });
+        saveAnnotation(
+          requestid,
+          astr,
+          (data) => {
+            fetchPageFlag(requestid, currentLayer.redactionlayerid, (error) =>
+              console.log(error)
+            );
+          },
+          (error) => {
+            console.log(error);
+          },
+          currentLayer.redactionlayerid,
+          createPageFlagPayload(
+            pageFlagSelections,
+            currentLayer.redactionlayerid
+          )
+        );
         annotManager.addAnnotations(sectionAnnotations);
+
         // Always redraw annotation
         sectionAnnotations.forEach((a) => annotManager.redrawAnnotation(a));
+        setNewRedaction(null);
       }
     };
 
@@ -1517,7 +1604,7 @@ const Redlining = React.forwardRef(
     useEffect(() => {
       while (deleteQueue?.length > 0) {
         let annot = deleteQueue.pop();
-        if (annot && annot.name !== newRedaction?.name) {
+        if (annot && !newRedaction?.names.includes(annot.name)) {
           let stitchedPageNo = Number(annot.page) + 1;
           let displayedDoc = pageMappedDocs.stitchedPageLookup[stitchedPageNo];
           deleteRedaction(
